[multimodal_gen] Rename k2.py modules to krea2.py to match Krea2* classes
The pipeline, transformer, configs, and stage classes are Krea2* (matching the
published model_index.json _class_name); rename their files k2.py -> krea2.py
and update the import paths so the filenames track the model family, as
elsewhere (qwen_image.py, flux.py, zimage.py). Pure rename -- class names,
registry keys, and generated output are unchanged.
