+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/model_specific_stages/krea2.py
